// packages/server/src/services/legacy-execution/backup-2025-06-05-160205/existing-run-infrastructure/types.ts
export type Location = {
    /** 
     * The ID of the location within the object.
     * Typically this is a node ID within a routine
     */
    locationId: Id;
    /** 
     * The ID of the object being run.
     * For example, if this location is for a node in a routine, this would be the routine ID 
     * (and not the node ID or subroutine ID).
     */
    objectId: Id;
    /** 
     * The type of object being run
     */
    objectType: `${ResourceType.Project | ResourceType.Routine}`;
    /**
     * The ID of the subroutine, if this location refers to a graph node that points to a subroutine.
     * Can also be the directory ID if running a project.
     */
    subroutineId: Id | null;
};
// ...
export type LocationStack = Location[];

export type Id = string;

/**
 * Behavior when a run reaches a limit.
 */
export type RunLimitBehavior = "Pause" | "Stop";
// ...
export type RunRequestLimits = {
    /** 
     * Maximum tokens in routine context, which accumulates relevant info 
     * over multiple steps. Having fewer tokens reduces the cost of 
     * generating LLM responses.
     */
    maxContextTokens?: number;
    /** Maximum number of credits that can be spent */
    maxCredits?: string;
    /** Maximum number of steps that can be run */
    maxSteps?: number;
    /** Maximum time that an be spent on the run, in milliseconds */
    maxTime?: number;
    /** What to do on max credits reached */
    onMaxCredits?: RunLimitBehavior;
    /** What to do in max time reached */
    onMaxTime?: RunLimitBehavior;
    /** What to do on max steps reached */
    onMaxSteps?: RunLimitBehavior;
}
// ...
export type RunDecisionConfig = {
    /** 
     * How subroutine inputs should be generated.
     * 
     * `Auto` - Let the AI generate the inputs
     * `Manual` - Let the user provide the inputs
     */
    inputGeneration: InputGenerationStrategy,
    /**
     * The type of path selection strategy to use for the run, 
     * when we're in a situation where there are multiple outgoing branches 
     * and we can only pick one or a subset of them.
     */
    pathSelection: PathSelectionStrategy,
    /** 
     * How subroutines should be executed.
     * 
     * `Auto` - Run the subroutine as soon as the inputs are generated
     * `Manual` - Make the user press a button to run the subroutine
     */
    subroutineExecution: SubroutineExecutionStrategy,
}
// ...
export type RunConfig = {
    /** Bot configuration for the run */
    botConfig: RunBotConfig;
    /**
     * The decision strategies to apply for the run.
     */
    decisionConfig: RunDecisionConfig;
    /**
     * Indicates if the run should be marked as private
     */
    isPrivate: boolean;
    /**
     * Indicates if the task is time-sensitive.
     * If so, the server may prioritize it over other tasks, 
     * at the cost of potentially higher credit usage.
     */
    isTimeSensitive?: boolean;
    /** Limits for the run */
    limits: RunRequestLimits,
    /** Config for the main loop that powers the run */
    loopConfig: RunUntilDoneConfig,
    /**
     * Run behavior when a branch fails for any reason 
     * (and there are more branches to run).
     * 
     * `Continue` - Continue running other branches as if nothing happened
     * `Pause` - Pause the run (meaning you'll have to manually resume it)
     * `Stop` (RECOMMENDED) - Stop the run and consider it failed
     */
    onBranchFailure?: "Continue" | "Pause" | "Stop",
    /**
     * Run behavior when a gateway fork has no active outgoing branches.
     * 
     * NOTE: This still applies for event-based gateways, since it's the subsequent nodes that handle 
     * the event - not the gateway itself.
     * 
     * `Continue` - Close the branch and continue the run
     * `Wait` - Wait for a valid branch to be available
     * `Fail` (RECOMMENDED) - Fail the run
     */
    onGatewayForkFailure?: "Continue" | "Wait" | "Fail",
    /**
     * Run behavior when a node (non-gateway and non-event) has no active outgoing branches.
     * 
     * `Continue` - Close the branch and continue the run
     * `Wait` - Wait for a valid branch to be available
     * `Fail` (RECOMMENDED) - Fail the run
     */
    onNormalNodeFailure?: "Continue" | "Wait" | "Fail",
    /**
     * Run behavior when the only remaining branches for a routine are waiting at gateways, and 
     * at least one branch has already failed or reached the end of the graph.
     * 
     * NOTE: If at least one branch is waiting for an event, this will not apply.
     * 
     * `Continue` (RECOMMENDED) - Close waiting branches and continue the run
     * `Pause` - Pause the run (meaning you'll have to manually resume it)
     * `Stop` - Stop the run and consider it failed
     */
    onOnlyWaitingBranches?: "Continue" | "Pause" | "Stop",
    /**
     * If true, all subroutines generate dummy data instead of running.
     * Should cost 0 credits.
     */
    testMode?: boolean,
}

/**
 * Represents the accumulated stats for current run.
 */
export type RunMetrics = {
    /** The total complexity completed so far */
    complexityCompleted: number;
    /** The total complexity of the run */
    complexityTotal: number;
    /** 
     * How many credits have been spent so far.
     * 
     * NOTE: This is a string because credits are stored using BigInt, which cannot 
     * be sent through the API or websockets.
     */
    creditsSpent: string;
    /** When the run was started */
    startedAt: Date | null;
    /** How many steps (excluding the current step) have been run so far. */
    stepsRun: number;
    /** How much time (in milliseconds) has elapsed in all steps combined */
    timeElapsed: number;
}

export type RunProgressStep = Location & {
    /** The complexity of the step */
    complexity: number;
    /** How many times the user switched tabs while completing the step, if run manually */
    contextSwitches: number;
    /** DUMMY_ID until saved, then DB ID */
    id: Id;
    /** When the step was started */
    startedAt: Date;
    /** When the step was completed */
    completedAt?: Date;
    /** The name of the step */
    name: string;
    /** The ID of the graph node this step corresponds to */
    nodeId: Id;
    /** The order of the step in the run */
    order: number;
    /** The primary ID of the resource version that the subroutine is a part of (i.e. the resource version that contains the graph config) */
    resourceInId: Id;
    /** The primary ID of the resource version at this graph node (i.e. the subroutine) */
    resourceVersionId?: Id;
    /** The status of the step */
    status: RunStepStatus;
    /** How much time (in milliseconds) has elapsed in the step (exluding paused/scheduled time) */
    timeElapsed?: number;
}

/** Represents the current run state. */
export type RunProgress = {
    /** Store the version number for this shape for future compatibility */
    __version: string;
    /** All live branches in the run (including nested branches) */
    branches: BranchProgress[];
    /** The run configuration, including limits and settings */
    config: RunConfig;
    /**
     * Decisions that have been deferred or resolved. 
     * Required for pausing a resuming a run due to blocking decisions, 
     * such as waiting for a user to manually choose an option.
     */
    decisions: (DeferredDecisionData | ResolvedDecisionData)[];
    /** The name of the run. Typically derived from the object being run */
    name: string;
    /** 
     * Who owns the run.
     * Typically the same as the user that requested the run, but can be different 
     * (e.g. the run is assigned to a team instead of a user).
     */
    owner: { __typename: "User" | "Team", id: Id };
    /** The run ID */
    runId: Id;
    /** The ID of the main object being run (either a routine or project) */
    runOnObjectId: Id | null;
    /** The schedule associated with the run, if any */
    schedule: ScheduleShape | null;
    /** Previous step history, in chronological order */
    steps: RunProgressStep[];
    /** Accumulated metrics (cost, time, etc.) from active and closed branches */
    metrics: RunMetrics;
    /** The current overall run status */
    status: RunStatus;
    /** 
     * The context collected for each multi-step subroutine, keyed by `subroutineInstanceId`. 
     * 
     * We collect subcontexts here so that branches can share context with each other.
     * 
     * We use `subroutineInstanceId` instead of `subroutineId` because the same subroutine  
     * may be triggered multiple times in the same run, and we want to keep the contexts separate.
     */
    subcontexts: Record<string, SubroutineContext>;
}

/**
 * Describes why a run changed status. 
 * Can be used to provide more context to the user or to the AI.
 */
export enum RunStatusChangeReason {
    AutomaticPause = "AutomaticPause",
    AutomaticStop = "AutomaticStop",
    Completed = "Completed",
    Error = "Error",
    MaxCredits = "MaxCredits",
    MaxLoops = "MaxLoops",
    MaxSteps = "MaxSteps",
    MaxTime = "MaxTime",
    UserCanceled = "UserCanceled",
    UserPaused = "UserPaused",
}
// ...
export enum BranchStatus {
    /** The branch is currently running with no blockers. */
    Active = "Active",
    /** The branch has completed successfully. */
    Completed = "Completed",
    /** The branch has failed. */
    Failed = "Failed",
    /** The branch is waiting for some condition to be met before continuing. */
    Waiting = "Waiting",
}
// ...
export type BranchProgress = {
    /** ID to identify this branch. */
    branchId: Id;
    /** 
     * If this branch triggered a multi-step subroutine, this is the instance ID of the subroutine. 
     * This is needed to determine when the subroutine has completed.
     */
    childSubroutineInstanceId: Id | null;
    /** Locations that are not allowed to be visited again */
    closedLocations: Location[],
    /** The number of credits spent so far in this branch, as a stringified bigint. */
    creditsSpent: string;
    /** The current location in the run. */
    locationStack: LocationStack;
    /** 
     * If we're running in a manual subroutineExecution mode, this flag 
     * lets us know if we've manually pressed the "Run" button for the subroutine.
     * If not, we can't run the subroutine.
     */
    manualExecutionConfirmed: boolean;
    /** The start time for the current node, in milliseconds since epoch. */
    nodeStartTimeMs: number | null;
    /** 
     * When a branch is created at a start node, this ID is generated. 
     * All other branches created during the execution of the graph will have the same ID.
     * 
     * We call this a process ID because in BPMN there should only be one start node triggered per process ID.
     * When a process (think swim lane) has multiple start nodes, a decision should be made to pick one of them.
     */
    processId: Id;
    /** The branch's status */
    status: BranchStatus;
    /** The current run progress step ID linked to this branch */
    stepId: Id | null;
    /**
     * A composite ID of `${subroutineId}.${uniqueId}` to distinguish this instance of a subroutine from others.
     * 
     * If a subroutine is triggered multiple times in the same run, this allows us to 
     * distinguish between them.
     * 
     * NOTE: This is a composite ID so that we can grab the subroutine ID from it if needed.
     */
    subroutineInstanceId: string;
    /** Whether the branch is part of a navigator that supports parallel execution */
    supportsParallelExecution: boolean;
}
// ...
export type DeferredDecisionData = {
    /** The type of decision resolved state */
    __type: "Waiting";
    /** 
     * An ID that uniquely identifies this decision, based on the branch ID 
     * and where in the state machine the decision is for.
     */
    key: string;
    /** A message to display to the user, along with the decision options */
    message?: string;
    /** The possible options for the decision */
    options: DecisionOption[];
    /** The type of decision to make */
    decisionType: "chooseOne" | "chooseMultiple";
}
// ...
export type ResolvedDecisionData = ResolvedDecisionDataChooseOne | ResolvedDecisionDataChooseMultiple;